Okay there is a lot going on here. First, I applied the principle of using real types as much as possible instead of strings. For example, there were many places where a string type name was used when Type or WichType or something with real type information should be used instead. The goal is to avoid converting things to strings as much as possible in the code generator, leaving all of that for the templates. You will notice that I made a model object for each type so now the types are explicitly converted to text:
IntType(t) ::= "int"
FloatType(t) ::= "double"
VectorType(t) ::= "Vector *"
StringType(t) ::= "String *"

As it was, float could only go to float because that's what its name was in the original language. But, a translator should be able to convert that to any other type specification in the target language.

I have also converted the return type of functions and type of temporary variables to model elements that are converted to templates. Otherwise we are once again forced to use a Wich symbol as an output type in the target language.

I have altered the grammar so that it has labels for each alternative in the type specification rule. This makes it very easy to have a method return the appropriate type model object in the code generator visitor:

	public OutputModelObject visitIntTypeSpec(WichParser.IntTypeSpecContext ctx) {
		return new IntType();
	}

I noticed that you did not have a visitor method for rule type, which indicates that we were not really performing a translation for types.

You'll also notice that I have added or subtracted white space to be consistent.

+++ trans/test/t13_expected.c
@@ -2,13 +2,13 @@
 #include "wich.h"
 
 String * f();
-float g();
+double g();
 
 String * f()
 {
 	g();
 }
-float g()
+double g()
 {
 	f();
 }
